Make page not have nested containers by default.
Previous behavior is fully supported if explicitly requested (when calling page and and content functions).
Having nested containers is a bad practice according to Bootstrap 3, and results in excessive left/right margins.
Re-order methods in PageContext to be more consistent with expected usage order (based on examples).
Allow the content to specify the ContainerWidth instead of the page.